refactor: re-organize core/nitro directory
