+++ packages/nuxt/src/core/runtime/nitro/handlers/error.ts
