+++ packages/nuxt/src/core/runtime/nitro/utils/paths.ts
